+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 2.0.1-SNAPSHOT
+- Adds `ngram` function
 
 ## 2.0.0 (March 19, 2018)
 - Adds support for @query type (user defined functions)